move code tidbreader code to tidb_reader
